refactor(js-sdk): self-cleaning tar stream for template upload (match Python)
Replace tarFileToTempFile ({ path, size, cleanup }) with tarFileToStream
({ stream, size }), mirroring the Python SDK's tar_file_stream: the
spooled archive deletes itself once the returned stream is closed
(consumed, errored, or destroyed) via its `close` event.

uploadFile no longer manages cleanup — there is no cleanup callback or
`finally` in the upload path, so a cleanup failure can no longer mask the
upload result. To preserve guaranteed removal when fetch never consumes
the body (e.g. it throws before reading), uploadFile destroys the stream
in its catch, which triggers the same `close`-based cleanup.

Node has no portable delete-on-close (Python's TemporaryFile uses
O_TEMPORARY on Windows / immediate unlink on POSIX), so cleanup is tied to
the stream lifecycle instead. Tests rewritten to consume the stream and
assert the spooled archive is removed on both consume and destroy paths.
