+++ packages/js-sdk/src/template/buildApi.ts
@@ -1,10 +1,11 @@
+import type { Readable } from 'node:stream'
 import { ApiClient, handleApiError, components } from '../api'
 import { buildRequestSignal } from '../connectionConfig'
 import { dynamicImport } from '../utils'
 import { BuildError, FileUploadError, TemplateError } from '../errors'
 import { FILE_UPLOAD_TIMEOUT_MS } from './consts'
 import { LogEntry } from './logger'
-import { getBuildStepIndex, tarFileToTempFile } from './utils'
+import { getBuildStepIndex, tarFileToStream } from './utils'
 import {
   BuildStatusReason,
   TemplateBuildStatus,
@@ -121,62 +122,52 @@ export async function uploadFile(
 ) {
   const { fileName, url, fileContextPath, ignorePatterns, resolveSymlinks } =
     options
+  // Spool the archive to a temporary file and stream it from disk instead of
+  // buffering in memory. S3 presigned PUT URLs reject Transfer-Encoding:
+  // chunked with 501 NotImplemented (see e2b-dev/e2b#1243), so the upload
+  // sends an explicit Content-Length, which fetch honors for stream bodies.
+  // The spooled file deletes itself once the stream is consumed, so there is
+  // no cleanup to manage here. The Python SDK takes the same approach
+  // (build_api.py:upload_file).
+  let stream: Readable | undefined
   try {
-    // Spool the archive to a temporary file so it can be streamed from disk
-    // instead of buffered in memory. S3 presigned PUT URLs reject
-    // Transfer-Encoding: chunked with 501 NotImplemented (see
-    // e2b-dev/e2b#1243), so the upload sends an explicit Content-Length,
-    // which fetch honors for stream bodies. The Python SDK takes the same
-    // approach (build_api.py:upload_file).
-    const {
-      path: tarPath,
-      size,
-      cleanup,
-    } = await tarFileToTempFile(
+    // Dynamically import so the browser bundle doesn't pull in node:stream.
+    const { Readable } =
+      await dynamicImport<typeof import('node:stream')>('node:stream')
+
+    const tar = await tarFileToStream(
       fileName,
       fileContextPath,
       ignorePatterns,
       resolveSymlinks
     )
+    stream = tar.stream
 
-    try {
-      // Dynamically import so the browser bundle doesn't pull in node:fs
-      // and node:stream.
-      const { createReadStream } =
-        await dynamicImport<typeof import('node:fs')>('node:fs')
-      const { Readable } =
-        await dynamicImport<typeof import('node:stream')>('node:stream')
+    const res = await fetch(url, {
+      method: 'PUT',
+      body: Readable.toWeb(stream) as ReadableStream<Uint8Array>,
+      headers: {
+        'Content-Length': tar.size.toString(),
+      },
+      // Streaming request bodies require half-duplex mode.
+      duplex: 'half',
+      signal: buildRequestSignal(
+        abortOpts?.requestTimeoutMs ?? FILE_UPLOAD_TIMEOUT_MS,
+        abortOpts?.signal
+      ),
+    } as RequestInit)
 
-      const res = await fetch(url, {
-        method: 'PUT',
-        body: Readable.toWeb(
-          createReadStream(tarPath)
-        ) as ReadableStream<Uint8Array>,
-        headers: {
-          'Content-Length': size.toString(),
-        },
-        // Streaming request bodies require half-duplex mode.
-        duplex: 'half',
-        signal: buildRequestSignal(
-          abortOpts?.requestTimeoutMs ?? FILE_UPLOAD_TIMEOUT_MS,
-          abortOpts?.signal
-        ),
-      } as RequestInit)
-
-      if (!res.ok) {
-        throw new FileUploadError(
-          `Failed to upload file: ${res.statusText}`,
-          stackTrace
-        )
-      }
-    } finally {
-      // Cleanup is best-effort: a temp-removal failure (e.g. the read
-      // stream still holds the file) must not mask a successful upload as
-      // a FileUploadError, nor overwrite a real upload error. A leaked
-      // temp dir is non-fatal — the OS reclaims it.
-      await cleanup().catch(() => {})
+    if (!res.ok) {
+      throw new FileUploadError(
+        `Failed to upload file: ${res.statusText}`,
+        stackTrace
+      )
     }
   } catch (error) {
+    // Ensure the spooled archive is removed even if fetch never consumed the
+    // stream (e.g. it threw before reading the body). Destroying the stream
+    // fires its `close` handler, which deletes the temp file.
+    stream?.destroy()
     if (error instanceof FileUploadError) {
       throw error
     }
